"fixed int division"

+++ BlinkLibrary/BCM/blinkBcm2835.c
@@ -2,9 +2,10 @@
 #include <bcm2835.h>
 #include <unistd.h>
 #include <gpiotopin.h>
+//gcc -o blinkBcm2835 blinkBcm2835.c -lbcm2835 for compile
 
 int main(){
-	int freq = 0;
+	double freq = 0;
 	printf("Enter frequency: ");
 	scanf("%d", &freq); //read int for freq
